Add some exception handling and OpenAPI descriptions.

+++ src/main/java/com/alexereh/messenger/config/ApplicationConfig.java
@@ -37,7 +37,7 @@ public class ApplicationConfig {
 	}
 
 	@Bean
-	public AuditorAware<Long> auditorAware() {
+	public AuditorAware<Integer> auditorAware() {
 		return new ApplicationAuditAware();
 	}
